refactor: SequentialTaskSet uses same class meta, refactored into task.py
